fix linux64

+++ src/oak/util/Stack.d
@@ -110,8 +110,8 @@ struct Stack (V, int Size = 0,  size_t Step = 64 ) {
                 assert(false, to!string(line) ) ;
         }
 
-        ptrdiff_t opApply (scope ptrdiff_t delegate(ref V value) dg) {
-		ptrdiff_t result;
+        int opApply (scope int delegate(ref V value) dg) {
+		int result;
 
 		for (ptrdiff_t i=depth; i-- && result is 0;){
 		     result = dg (stack[i]) ;
